fix warning

+++ src/lib/jp2/t1/t1_ht/T1HT.cpp
@@ -147,7 +147,7 @@ bool T1HT::decode(decodeBlockInfo *block) {
 	}
 	size_t offset = 0;
 	// note: min_buf_vec only contains segments of non-zero length
-	for (int32_t i = 0; i < min_buf_vec->size(); ++i) {
+	for (size_t i = 0; i < min_buf_vec->size(); ++i) {
 		grk_buf *seg = (grk_buf*) min_buf_vec->get(i);
 		memcpy(coded_data + offset, seg->buf, seg->len);
 		offset += seg->len;
